requirements_privacy_preserving.txt
numpy>=1.21.0
opencv-python>=4.5.0
pandas>=1.3.0
matplotlib>=3.5.0
pillow>=8.0.0
scipy>=1.7.0
tqdm>=4.62.0 